requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
openpyxl
requests
great-expectations
faker
kafka-python